+++ examples-pydantic-v1/omni_pydantic_v1/data_classes/_core.py
@@ -30,14 +30,20 @@ from typing import (
 import pandas as pd
 from cognite.client import CogniteClient
 from cognite.client import data_modeling as dm
-from cognite.client.data_classes import TimeSeriesList
+from cognite.client.data_classes import TimeSeriesList, Datapoints
 from cognite.client.data_classes.data_modeling.instances import (
     Instance,
     InstanceApply,
     Properties,
     PropertyValue,
 )
+from cognite.client.utils import datetime_to_ms
 from pydantic import BaseModel, Extra, Field, root_validator
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    from typing_extensions import Self
 
 
 DEFAULT_QUERY_LIMIT = 5
@@ -52,6 +58,43 @@ class _NotSetSentinel:
     It is used when we need to distinguish between not set and None."""
 
     ...
+
+
+class TimeSeriesGraphQL(BaseModel):
+    class Config:
+        arbitrary_types_allowed = True
+        allow_population_by_field_name = True
+
+    id: Optional[int] = None
+    external_id: Optional[str] = Field(None, alias="externalId")
+    instance_id: Optional[dm.NodeId] = Field(None, alias="instanceId")
+    name: Optional[str] = None
+    is_string: Optional[bool] = Field(None, alias="isString")
+    metadata: Optional[dict[str, str]] = None
+    unit: Optional[str] = None
+    unit_external_id: Optional[str] = Field(None, alias="unitExternalId")
+    asset_id: Optional[int] = Field(None, alias="assetId")
+    is_step: Optional[bool] = Field(None, alias="isStep")
+    description: Optional[str] = None
+    security_categories: Optional[list[int]] = Field(None, alias="securityCategories")
+    data_set_id: Optional[int] = Field(None, alias="dataSetId")
+    created_time: Optional[int] = Field(None, alias="createdTime")
+    last_updated_time: Optional[int] = Field(None, alias="lastUpdatedTime")
+    data: Optional[Datapoints] = None
+
+    @root_validator(pre=True, allow_reuse=True)
+    def parse_datapoints(cls, data: Any) -> Any:
+        if isinstance(data, dict) and "getDataPoints" in data:
+            datapoints = data.pop("getDataPoints")
+            if "items" in datapoints:
+                for item in datapoints["items"]:
+                    # The Datapoints expects the timestamp to be in milliseconds
+                    item["timestamp"] = datetime_to_ms(
+                        datetime.datetime.fromisoformat(item["timestamp"].replace("Z", "+00:00"))
+                    )
+                data["datapoints"] = datapoints["items"]
+                data["data"] = Datapoints.load(data)
+        return data
 
 
 @dataclass