+++ examples-pydantic-v1/scenario_instance_pydantic_v1/client/data_classes/__init__.py
@@ -12,6 +12,7 @@ from ._core import (
     ResourcesWrite,
     ResourcesWriteResult,
     PageInfo,
+    TimeSeriesGraphQL,
 )
 from ._scenario_instance import (
     ScenarioInstance,
@@ -23,6 +24,11 @@ from ._scenario_instance import (
     ScenarioInstanceTextFields,
     ScenarioInstanceWrite,
     ScenarioInstanceWriteList,
+)
+
+
+ScenarioInstanceGraphQL.update_forward_refs(
+    TimeSeriesGraphQL=TimeSeriesGraphQL,
 )
 
 
@@ -40,6 +46,7 @@ __all__ = [
     "GraphQLList",
     "ResourcesWriteResult",
     "PageInfo",
+    "TimeSeriesGraphQL",
     "ScenarioInstance",
     "ScenarioInstanceGraphQL",
     "ScenarioInstanceWrite",