+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_generator.py
@@ -67,8 +67,8 @@ class GeneratorGraphQL(GraphQLCore):
     """
 
     view_id: ClassVar[dm.ViewId] = dm.ViewId("power-models", "Generator", "1")
-    generator_speed_controller: Union[TimeSeries, dict, None] = None
-    generator_speed_controller_reference: Union[TimeSeries, dict, None] = None
+    generator_speed_controller: Union[TimeSeriesGraphQL, dict, None] = None
+    generator_speed_controller_reference: Union[TimeSeriesGraphQL, dict, None] = None
 
     @root_validator(pre=True)
     def parse_data_record(cls, values: Any) -> Any: