+++ examples-pydantic-v1/windmill_pydantic_v1/data_classes/_metmast.py
@@ -73,9 +73,9 @@ class MetmastGraphQL(GraphQLCore):
 
     view_id: ClassVar[dm.ViewId] = dm.ViewId("power-models", "Metmast", "1")
     position: Optional[float] = None
-    temperature: Union[TimeSeries, dict, None] = None
-    tilt_angle: Union[TimeSeries, dict, None] = None
-    wind_speed: Union[TimeSeries, dict, None] = None
+    temperature: Union[TimeSeriesGraphQL, dict, None] = None
+    tilt_angle: Union[TimeSeriesGraphQL, dict, None] = None
+    wind_speed: Union[TimeSeriesGraphQL, dict, None] = None
 
     @root_validator(pre=True)
     def parse_data_record(cls, values: Any) -> Any: