+++ examples/equipment_unit/data_classes/__init__.py
@@ -12,6 +12,7 @@ from ._core import (
     ResourcesWrite,
     ResourcesWriteResult,
     PageInfo,
+    TimeSeriesGraphQL,
 )
 from ._equipment_module import (
     EquipmentModule,
@@ -57,6 +58,10 @@ from ._work_order import (
     WorkOrderWriteList,
 )
 
+EquipmentModule.model_rebuild()
+EquipmentModuleGraphQL.model_rebuild()
+EquipmentModuleWrite.model_rebuild()
+EquipmentModuleApply.model_rebuild()
 UnitProcedure.model_rebuild()
 UnitProcedureGraphQL.model_rebuild()
 UnitProcedureWrite.model_rebuild()
@@ -81,6 +86,7 @@ __all__ = [
     "GraphQLList",
     "ResourcesWriteResult",
     "PageInfo",
+    "TimeSeriesGraphQL",
     "EquipmentModule",
     "EquipmentModuleGraphQL",
     "EquipmentModuleWrite",