+++ tests/test_integration/conftest.py
@@ -14,9 +14,11 @@ from tests.constants import IS_PYDANTIC_V2, OMNI_SDK
 
 if IS_PYDANTIC_V2:
     from omni import OmniClient
+    from scenario_instance.client import ScenarioInstanceClient
     from windmill import WindmillClient
 else:
     from omni_pydantic_v1 import OmniClient
+    from scenario_instance_pydantic_v1.client import ScenarioInstanceClient
     from windmill_pydantic_v1 import WindmillClient
 
 
@@ -79,6 +81,11 @@ def omni_client(cognite_client: CogniteClient) -> OmniClient:
 @pytest.fixture(scope="session")
 def wind_client(cognite_client: CogniteClient) -> WindmillClient:
     return WindmillClient(cognite_client)
+
+
+@pytest.fixture(scope="session")
+def scenario_instance_client(cognite_client: CogniteClient) -> ScenarioInstanceClient:
+    return ScenarioInstanceClient(cognite_client)
 
 
 @pytest.fixture(scope="session")